feat: Improve Error handling and Input Validation (#9)

+++ orders-service/internal/firebase/customer.go
@@ -7,6 +7,8 @@ import (
 	"cloud.google.com/go/firestore"
 	"github.com/Mik3y-F/order-management-system/orders/internal/service"
 	"google.golang.org/api/iterator"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
 var _ service.CustomerService = (*CustomerService)(nil)
@@ -40,11 +42,15 @@ func (s *CustomerService) CreateCustomer(ctx context.Context, customer *service.
 	customer.CreatedAt = currentTime.Format(time.RFC3339)
 	customer.UpdatedAt = currentTime.Format(time.RFC3339)
 
+	err := customer.Validate()
+	if err != nil {
+		return nil, service.Errorf(service.INVALID_ERROR, "invalid customer provided: %v", err)
+	}
 	customerModel := s.marshallCustomer(customer)
 
-	docRef, _, writeErr := s.customerCollection().Add(ctx, customerModel)
-	if writeErr != nil {
-		return nil, writeErr
+	docRef, _, err := s.customerCollection().Add(ctx, customerModel)
+	if err != nil {
+		return nil, service.Errorf(service.INTERNAL_ERROR, "failed to create customer: %v", err)
 	}
 
 	customer.Id = docRef.ID
@@ -56,14 +62,20 @@ func (s *CustomerService) GetCustomer(ctx context.Context, id string) (*service.
 
 	s.CheckPreconditions()
 
-	docRef, getErr := s.customerCollection().Doc(id).Get(ctx)
-	if getErr != nil {
-		return nil, getErr
+	if id == "" {
+		return nil, service.Errorf(service.INVALID_ERROR, "id is required")
+	}
+
+	docRef, err := s.customerCollection().Doc(id).Get(ctx)
+	if status.Code(err) == codes.NotFound {
+		return nil, service.Errorf(service.NOT_FOUND_ERROR, "customer not found")
+	} else if err != nil {
+		return nil, service.Errorf(service.INTERNAL_ERROR, "failed to get customer: %v", err)
 	}
 
 	customerModel := &CustomerModel{}
 	if err := docRef.DataTo(customerModel); err != nil {
-		return nil, err
+		return nil, service.Errorf(service.INTERNAL_ERROR, "failed to unmarshall customer: %v", err)
 	}
 
 	customer := s.unmarshallCustomer(customerModel)
@@ -86,12 +98,12 @@ func (s *CustomerService) ListCustomers(ctx context.Context) ([]*service.Custome
 			break
 		}
 		if err != nil {
-			return nil, err
+			return nil, service.Errorf(service.INTERNAL_ERROR, "failed to iterate customers: %v", err)
 		}
 
 		customerModel := &CustomerModel{}
 		if err := docRef.DataTo(customerModel); err != nil {
-			return nil, err
+			return nil, service.Errorf(service.INTERNAL_ERROR, "failed to unmarshall customer: %v", err)
 		}
 
 		customer := s.unmarshallCustomer(customerModel)
@@ -108,28 +120,52 @@ func (s *CustomerService) UpdateCustomer(
 
 	s.CheckPreconditions()
 
-	customer, getErr := s.GetCustomer(ctx, id)
-	if getErr != nil {
-		return nil, getErr
+	customer, err := s.GetCustomer(ctx, id)
+	if err != nil {
+		return nil, err
 	}
 
-	// todo: update only the fields that are not empty
-	customer.FirstName = update.FirstName
-	customer.LastName = update.LastName
-	customer.Email = update.Email
+	if c := update.FirstName; c != nil {
+		customer.FirstName = *c
+	}
+
+	if c := update.LastName; c != nil {
+		customer.LastName = *c
+	}
+
+	if c := update.Email; c != nil {
+		customer.Email = *c
+	}
+
+	if c := update.Phone; c != nil {
+		customer.Phone = *c
+	}
 
 	timeNow := time.Now()
 	customer.UpdatedAt = timeNow.Format(time.RFC3339)
 
+	err = customer.Validate()
+	if err != nil {
+		return nil, service.Errorf(service.INVALID_ERROR, "invalid customer details provided: %v", err)
+	}
+
 	customerModel := s.marshallCustomer(customer)
 
-	_, writeErr := s.customerCollection().Doc(id).Set(ctx, customerModel)
+	_, err = s.customerCollection().Doc(id).Set(ctx, customerModel)
+	if err != nil {
+		return nil, service.Errorf(service.INTERNAL_ERROR, "failed to update customer: %v", err)
+	}
 
-	return customer, writeErr
+	return customer, nil
 
 }
 
 func (s *CustomerService) DeleteCustomer(ctx context.Context, id string) error {
+	s.CheckPreconditions()
+
+	if id == "" {
+		return service.Errorf(service.INVALID_ERROR, "id is required")
+	}
 
 	_, err := s.customerCollection().Doc(id).Delete(ctx)
 
@@ -142,6 +178,7 @@ func (s *CustomerService) marshallCustomer(customer *service.Customer) *Customer
 		FirstName: customer.FirstName,
 		LastName:  customer.LastName,
 		Email:     customer.Email,
+		Phone:     customer.Phone,
 		CreatedAt: customer.CreatedAt,
 		UpdatedAt: customer.UpdatedAt,
 	}
@@ -153,6 +190,7 @@ func (s *CustomerService) unmarshallCustomer(customerModel *CustomerModel) *serv
 		FirstName: customerModel.FirstName,
 		LastName:  customerModel.LastName,
 		Email:     customerModel.Email,
+		Phone:     customerModel.Phone,
 		CreatedAt: customerModel.CreatedAt,
 		UpdatedAt: customerModel.UpdatedAt,
 	}
